Polishing #2769
Polishing "Pipeline tests for Json and Search Modules" (f8af975289ac89af13c0d3d5f03b446d42eaa11c)
